refactor: move monitor into monitor.go
