+++ host/registry_test.go
@@ -118,104 +118,6 @@ func TestRegistryStartStop(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func TestGetSliceDiffs(t *testing.T) {
-	t.Parallel()
-	cases := []struct {
-		name           string
-		old, next      []string
-		expectedAdds   []string
-		expectedRmoves []string
-	}{
-		{
-			name: "both empty",
-			old:  []string{},
-			next: []string{},
-		},
-		{
-			name: "both nil",
-		},
-		{
-			name:         "old empty, all added",
-			old:          []string{},
-			next:         []string{"bar", "foo"},
-			expectedAdds: []string{"bar", "foo"},
-		},
-		{
-			name:           "next empty, all removed",
-			old:            []string{"bar", "foo"},
-			next:           []string{},
-			expectedRmoves: []string{"bar", "foo"},
-		},
-		{
-			name:         "add one",
-			old:          []string{"foo", "bar"},
-			next:         []string{"foo", "bar", "baz"},
-			expectedAdds: []string{"baz"},
-		},
-		{
-			name: "no change",
-			old:  []string{"foo", "bar"},
-			next: []string{"foo", "bar"},
-		},
-		{
-			name:           "remove one",
-			old:            []string{"bar", "baz", "foo"},
-			next:           []string{"bar", "foo"},
-			expectedRmoves: []string{"baz"},
-		},
-		{
-			name:           "adds and removals",
-			old:            []string{"bar", "foo"},
-			next:           []string{"baz", "qux"},
-			expectedAdds:   []string{"baz", "qux"},
-			expectedRmoves: []string{"bar", "foo"},
-		},
-		{
-			name:         "add in front",
-			old:          []string{"foo"},
-			next:         []string{"bar", "baz", "foo"},
-			expectedAdds: []string{"bar", "baz"},
-		},
-		{
-			name:         "add at back",
-			old:          []string{"bar"},
-			next:         []string{"bar", "foo", "qux"},
-			expectedAdds: []string{"foo", "qux"},
-		},
-		{
-			name:           "remove from front",
-			old:            []string{"bar", "baz", "foo"},
-			next:           []string{"foo"},
-			expectedRmoves: []string{"bar", "baz"},
-		},
-		{
-			name:           "remove from back",
-			old:            []string{"bar", "foo", "qux"},
-			next:           []string{"bar"},
-			expectedRmoves: []string{"foo", "qux"},
-		},
-		{
-			name:           "mixed adds, removes, and unchanged",
-			old:            []string{"alpha", "bravo", "delta", "echo", "golf", "hotel"},
-			next:           []string{"alpha", "charlie", "delta", "foxtrot", "golf", "india"},
-			expectedAdds:   []string{"charlie", "foxtrot", "india"},
-			expectedRmoves: []string{"bravo", "echo", "hotel"},
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			t.Parallel()
-			slices.Sort(c.old)
-			slices.Sort(c.next)
-			adds, removes := getSliceDiffs(c.old, c.next)
-
-			require.Equal(t, c.expectedAdds, adds)
-			require.Equal(t, c.expectedRmoves, removes)
-		})
-	}
-}
-
 func TestDeregisterUnknownHost(t *testing.T) {
 	t.Parallel()
 	logger, _ := zap.NewDevelopment()
